pyproject.toml
[tool.poetry]
name = "ckpt"
version = "0.1.0"
description = "Checkpointing functions for easier debugging"
authors = ["Holger Hoefling <hhoeflin@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
GitPython = "^3.1.27"
typer = "^0.4.0"
ipdb = {version = "^0.13.9", optional = true}

[tool.poetry.dev-dependencies]
pytest = "^7.1.1"
pytest-pudb = "^0.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
checkpoint = "ckpt.main:app"

[tool.black]
line-length = 88

[tool.pyright]
include = ["checkpoint", "tests"]
